// packages/stack-orval/src/generated/model/getPostsApiParams.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * MyApp API
 * ユーザー管理API
 * OpenAPI spec version: 1.0.0
 */

export type GetPostsApiParams = {
/**
 * ページ番号（1から開始）
 */
page?: string;
/**
 * 1ページあたりの件数（最大100）
 */
limit?: string;
/**
 * ユーザーIDでフィルタリング
 */
userId?: string;
/**
 * 公開状態でフィルタリング
 */
published?: string;
};